Switched to a newer, more compact neural network implementation.
- Now using doubles everywhere.
- Updated the gene file to a new format.
- Tweaked the mutations.
- Change serialization to use streams instead of the C api.

+++ StrangeWorld/MathAccel.cpp
@@ -5,5 +5,5 @@ double* MathAccel::cosLookup = NULL;
 double* MathAccel::sinLookup = NULL;
 double* MathAccel::toRadLookup = NULL;
 #if defined( USE_SQRT_LOOKUP )
-int* MathAccel::sqrtTableLookup = NULL;
+double* MathAccel::sqrtTableLookup = NULL;
 #endif
